Create project_description
Description of the project:
The objectif behind this project is to allow volunteers to participate in some sofware developments projects. Every project has taks and beneficiaires.
A Task has resources and each resource with a required number and a title (skill). For exemple the resources needed for a specific task can be
One Java Architects, two Junior Java developers, One scrum master. 
The system must have two kind of users, a volunteer and an administrator. A volunteer must have a title which indeed is his skill, scrum
master, java developer....

Here are Additional requirements


    1)We Should be able to:
        1.create a project with a description, location, and timeframe (expected start and end dates)
        Manage (add, delete, update) the tasks required to be done in a project. Each task should have a timeframe, status of the tasks, indicate what type of resources are required for each task (i.e, what skill if it is a volunteer or an asset that is required; include how many resources of that type are required).
        2.The projects may have a list of beneficiaries.
        3.Should be able to load html files to enhance the description of the project and beneficiaries information.
    2)Volunteers should be able to offer their services for tasks on projects (satisfy the need of a resource).
    3)List information about projects and their beneficiaries
    4)List tasks for a project
    5)List projects by status
    6)ook for projects that requires a particular type of resource
    7)Search projects by keywords and location
    8)List projects and tasks where a volunteer have offered services, ordered by date of the task.